requirements.txt
chromadb
hdbscan
langchain
langchain_community
langchain_ollama
llama-index
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-vector-stores-chroma
matplotlib
nf-core
plotly
seaborn
umap-learn